Add explicit allowed-origins options for '*' and matching any origin
This change is necessary to use credentials with any origin. Using
`Access-Control-Allow-Credentials: true` and
`Access-Control-Allow-Origin: *` is not valid per the spec. To permit
credentials, the exact origin needs to be returned in
`Access-Control-Allow-Origin`. To acheive this we've added the
`:match-origin` option which will always return the provided Origin
header. To make things more explicit, an option `:star-origin` was added
which matches the current behaviour of `nil`. `nil` is no longer a valid
option for `:allowed-origins`.